Only set componentProps.onLongPress if this.props.onLongPress is set
Setting `onLongPress` makes the whole component "pressable", even if the callback does nothing. "Pressable" components will capture press gestures, which means any ancestor components that might want to capture press gestures will be unable to do so.

We should only set `onLongPress` if we have a callback to call, so that any ancestor components that might want to capture a press gesture can do so.